LCA/requirements.txt
# ===============================
# Love Compatibility Analyzer
# Requirements File
# ===============================
# Backend & Web Framework
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==3.0.4

# Database & ORM
SQLAlchemy==2.0.31

# Machine Learning & Data Processing
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2

# AI & API Integration
requests==2.32.3
google-generativeai==0.8.2   # Gemini API wrapper

# Authentication & Security
bcrypt==4.2.0
python-dotenv==1.0.1

# Visualization & Frontend Integration
jinja2==3.1.4
chartjs==0.5.0                # Untuk integrasi Chart.js di dashboard (opsional)

# Testing & Development Tools
pytest==8.3.3
pytest-flask==1.3.0
coverage==7.6.1

# Deployment
gunicorn==23.0.0
gevent==24.2.1

# Utilities
itsdangerous==2.2.0
email-validator==2.2.0
